+++ src/network/index.js
@@ -1,0 +1,3 @@
+// network/index.js
+export {startNetwork} from './networkManager.js';
+export {loadKnownPeers, saveKnownPeers} from './utils/manageKnownPeers.js';
